pyproject.toml
[project]
name = "selfindex"
version = "0.1.0"
description = "Personal RAG workspace for archived conversations and future LLM-powered APIs."
requires-python = ">=3.10,<3.14"
dependencies = [
    "chromadb>=1.5.5",
    "flask>=3.1.3",
    "ijson>=3.5.0",
    "markdown>=3.10.2",
    "python-dotenv>=1.2.2",
    "torch>=2.10.0",
]
